Create a new hash object for each trap
Previously, the code was reusing the same hash object every time there
was a new trap. If the Hash is modified on the subsequent plugins, the
modification gets preserved for the next trap. That's undesired
behavior.

With this change, a new Hash object gets created every time a new trap
comes to fluentd. So, any modification to it will not affect other
events.